🖥用户界面调用接口并显示信息
