+++ src/views/Home.js
@@ -8,23 +8,25 @@ import ClippedDrawer from "../components/Drawer";
 class Home extends React.Component {
   render() {
     return (
-      <Container component="main" maxWidth="md">
+      <>
         <ClippedDrawer />
-        <Stack
-          direction={{ xs: "column", sm: "row" }}
-          spacing={{ xs: 0, sm: 1, md: 2 }}
-          sx={{
-            marginTop: 12,
-            display: "flex",
-            flexDirection: "column",
-            alignItems: "center",
-          }}
-        >
-          <ActionAreaCard gamename="2048" rating={4} introduction="2048小游戏" />
-          <ActionAreaCard gamename="PVZ" rating={3.5} />
-          <ActionAreaCard gamename="LOLM" rating={3.8} />
-        </Stack>
-      </Container>
+        <Container component="main" maxWidth="md">
+          <Stack
+            direction={{ xs: "column", sm: "row" }}
+            spacing={{ xs: 0, sm: 1, md: 2 }}
+            sx={{
+              marginTop: 12,
+              display: "flex",
+              flexDirection: "column",
+              alignItems: "center",
+            }}
+          >
+            <ActionAreaCard gamename="2048" rating={4} introduction="2048小游戏" />
+            <ActionAreaCard gamename="PVZ" rating={3.5} />
+            <ActionAreaCard gamename="LOLM" rating={3.8} />
+          </Stack>
+        </Container>
+      </>
     );
   }
 }
